feat: lazy load tabs and speed up analytics
