+++ src/lib/analytics.js
@@ -1,6 +1,6 @@
 export function startOfISOWeek(d) {
   const date = new Date(d);
-  const day = (date.getDay() + 6) % 7; // 0=Monday
+  const day = (date.getDay() + 6) % 7;
   date.setDate(date.getDate() - day);
   const weekNumber = (() => {
     const tmp = new Date(date.getTime());
@@ -37,12 +37,26 @@ export function validSet(s) {
   return s && s.mode !== 'time' && Number.isFinite(s.reps) && Number.isFinite(s.weightKg) && s.reps > 0 && s.weightKg >= 0;
 }
 
-export function freqDaysByGroup(sessions, repo, { from, to, routineFilter }) {
+const hashParams = (p) => JSON.stringify(p);
+const freqCache = new Map();
+const heatCache = new Map();
+const histCache = new Map();
+
+export function freqDaysByGroup(sessions, repo, { from, to, routineFilter } = {}) {
+  const key = hashParams({
+    from: from?.toISOString?.() || null,
+    to: to?.toISOString?.() || null,
+    routineFilter: routineFilter || null,
+    lenSessions: sessions.length,
+    lastSessionAt: sessions[sessions.length - 1]?.dateISO || sessions[sessions.length - 1]?.at || null,
+  });
+  if (freqCache.has(key)) return freqCache.get(key);
   const byGroupDay = new Map();
   for (const sess of sessions) {
     if (sess.type !== 'strength') continue;
     const d = new Date(sess.dateISO || sess.at || Date.now());
-    if (d < from || d > to) continue;
+    if (from && d < from) continue;
+    if (to && d > to) continue;
     if (routineFilter && sess.routineKey && sess.routineKey !== routineFilter) continue;
     const dk = dayKey(d);
     const seen = new Set();
@@ -57,17 +71,27 @@ export function freqDaysByGroup(sessions, repo, { from, to, routineFilter }) {
       byGroupDay.get(g).add(dk);
     }
   }
-  const out = Array.from(byGroupDay.entries()).map(([group, daysSet]) => ({ group, days: daysSet.size }));
+  const out = Array.from(byGroupDay.entries()).map(([group, daysSet]) => ({ group, days: daysSet.size })).filter((r) => Number.isFinite(r.days) && r.days > 0);
   out.sort((a, b) => b.days - a.days);
+  freqCache.set(key, out);
   return out;
 }
 
-export function heatmapWeekGroup(sessions, repo, { from, to, routineFilter }) {
+export function heatmapWeekGroup(sessions, repo, { from, to, routineFilter } = {}) {
+  const key = hashParams({
+    from: from?.toISOString?.() || null,
+    to: to?.toISOString?.() || null,
+    routineFilter: routineFilter || null,
+    lenSessions: sessions.length,
+    lastSessionAt: sessions[sessions.length - 1]?.dateISO || sessions[sessions.length - 1]?.at || null,
+  });
+  if (heatCache.has(key)) return heatCache.get(key);
   const matrix = new Map();
   for (const sess of sessions) {
     if (sess.type !== 'strength') continue;
     const d = new Date(sess.dateISO || sess.at || Date.now());
-    if (d < from || d > to) continue;
+    if (from && d < from) continue;
+    if (to && d > to) continue;
     if (routineFilter && sess.routineKey && sess.routineKey !== routineFilter) continue;
     const w = startOfISOWeek(d);
     const dk = dayKey(d);
@@ -75,9 +99,9 @@ export function heatmapWeekGroup(sessions, repo, { from, to, routineFilter }) {
     for (const st of sess.sets || []) {
       if (!validSet(st)) continue;
       const g = primaryGroupOf(st.exerciseId, repo, st.exerciseName);
-      const key = g + '|' + dk;
-      if (seen.has(key)) continue;
-      seen.add(key);
+      const key2 = g + '|' + dk;
+      if (seen.has(key2)) continue;
+      seen.add(key2);
       if (!matrix.has(w)) matrix.set(w, new Map());
       const row = matrix.get(w);
       row.set(g, (row.get(g) || 0) + 1);
@@ -90,12 +114,22 @@ export function heatmapWeekGroup(sessions, repo, { from, to, routineFilter }) {
   const values = {};
   for (const w of weeks) {
     values[w] = {};
-    for (const g of groups) values[w][g] = matrix.get(w).get(g) || 0;
+    for (const g of groups) {
+      const v = matrix.get(w).get(g) || 0;
+      values[w][g] = Number.isFinite(v) ? v : 0;
+    }
   }
-  return { weeks, groups, values };
+  const result = { weeks, groups, values };
+  heatCache.set(key, result);
+  return result;
 }
 
 export function buildPerExerciseHistory(sessions, exercisesById) {
+  const key = hashParams({
+    lenSessions: sessions.length,
+    lastSessionAt: sessions[sessions.length - 1]?.dateISO || sessions[sessions.length - 1]?.at || null,
+  });
+  if (histCache.has(key)) return histCache.get(key);
   const map = new Map();
   for (const sess of sessions) {
     if (sess.type !== 'strength') continue;
@@ -119,8 +153,12 @@ export function buildPerExerciseHistory(sessions, exercisesById) {
         byDay[it.date].oneRM = Math.max(byDay[it.date].oneRM, it.oneRM);
       }
     }
-    map.set(k, Object.values(byDay).sort((a, b) => a.date.localeCompare(b.date)));
+    const res = Object.values(byDay)
+      .filter((d) => Number.isFinite(d.volume) && Number.isFinite(d.oneRM))
+      .sort((a, b) => a.date.localeCompare(b.date));
+    map.set(k, res);
   }
+  histCache.set(key, map);
   return map;
 }
