+++ src/lib/repo.js
@@ -1,4 +1,4 @@
-import repo from '../data/exercisesRepo.json';
+import repo from '../data/exercisesRepo.json' assert { type: 'json' };
 
 /**
  * Load the exercises repository.
